feat(functional): dense tiled FDK backprojection — contraction reverse instead of scatter
The FDK bilinear gather's Enzyme reverse is an XLA scatter, which made the
pipeline gradient 7× its forward at 64/100 (the projector's own gradient is
1.5×). backproject_dense writes the same sampling as contractions: per view
batch and static image tile (one compiled loop over tiles), dense bilinear
weights over the rows and a per-(tile, view) column window (host table from
the tile's corner rays, static width), G = Σ_row wr·filt then Σ_k wc·G.

- fdk(...; dense = true, tile = 16) → backproject_dense; :gather stays the oracle
- pipelines: batching gains :tile; _auto_batching sizes the FDK view batch from
  the tile's window; the pipeline :projector knob governs DD and FDK together
- windowed projector: traced table rows in _window_stack2/_scalar_start
- probe design/reactant/probes/bench_fdk.jl

Host parity vs backproject: 1e-16 (Float64) / 3e-7 (Float32) over arc/flat,
narrow windows (22 of 256 columns), ragged tiles, loop and unrolled.

+++ src/functional/projection/dd_dense.jl
@@ -263,8 +263,8 @@ function dd_project_dense_windowed_run(vol::AbstractArray{<:Any, 4}, p::DDRunPla
 end
 
 # stack the per-view (transverse window × slab block) of V: Vw[:, :, :, b] = V[s_b : s_b+w-1, :, l0 : l0+blen-1]
-function _window_stack2(V, st, row::Int, w::Int, l0, blen::Int, B::Int)
+function _window_stack2(V, st, row, w::Int, l0, blen::Int, B::Int)
     parts = [reshape(_dslice_at(V, (_scalar_start(st, row, b), 1, l0), (w, size(V, 2), blen)), w, size(V, 2), blen, 1) for b in 1:B]
     return reduce((a, b) -> cat(a, b; dims = 4), parts)
 end
-_scalar_start(st, j::Int, b::Int) = Int(st[j, b])            # host; the extension returns a traced integer
+_scalar_start(st, j, b::Int) = Int(st[j, b])                 # host (j :: Int); the extension returns a traced integer
